Refactory the command folder.
