+++ src/app/registry.ts
@@ -4,15 +4,15 @@ import { ArbValidationRepository } from "./arb_validation/arb_validation.reposit
 import { ArbValidationService } from "./arb_validation/arb_validation.service";
 import { TranslationCacheDataSource } from "./cache/translation_cache.datasource";
 import { TranslationCacheRepository } from "./cache/translation_cache.repository";
-import { ConfigureTargetLanguageCodeCmd } from "./command/configure_target_language_code.cmd";
-import { CreateTranslationCacheCmd } from "./command/create_translation_cache.cmd";
-import { DecodeAllHtmlEntitiesCmd } from "./command/decode_all_html_entities.cmd";
-import { ExcludeTranslationCmd } from "./command/exclude_translation.cmd";
-import { InitializeCmd } from "./command/initialize.cmd";
-import { OpenGoogleSheetCmd } from "./command/open_google_sheet.cmd";
-import { TranslateCmd } from "./command/translate.cmd";
-import { UploadToGoogleSheetCmd } from "./command/upload_to_google_sheet.cmd";
-import { ValidateTranslationCmd } from "./command/validate_translation.cmd";
+import { ConfigureTargetLanguageCodeCmd } from "./command/configure/configure_target_language_code.cmd";
+import { ExcludeTranslationCmd } from "./command/configure/exclude_translation.cmd";
+import { InitializeCmd } from "./command/configure/initialize.cmd";
+import { OpenGoogleSheetCmd } from "./command/google_sheet/open_google_sheet.cmd";
+import { UploadToGoogleSheetCmd } from "./command/google_sheet/upload_to_google_sheet.cmd";
+import { CreateTranslationCacheCmd } from "./command/translate/create_translation_cache.cmd";
+import { TranslateCmd } from "./command/translate/translate.cmd";
+import { DecodeAllHtmlEntitiesCmd } from "./command/validate/decode_all_html_entities.cmd";
+import { ValidateTranslationCmd } from "./command/validate/validate_translation.cmd";
 import { ConfigRepository } from "./config/config.repository";
 import { ConfigService } from "./config/config.service";
 import { GoogleAuthService } from "./google_sheet/google_auth.service";
